requirements/requirements-dev.txt
flake8~=6.0
pep8-naming~=0.13.3
flake8-import-order~=0.18.2
flake8-quotes~=3.3.2

black~=23.1
djlint~=1.19.17
django-debug-toolbar~=4.0.0